+++ java/examples/datatypes/H5Ex_T_CompoundAttribute.java
@@ -54,15 +54,15 @@ public class H5Ex_T_CompoundAttribute {
         static int[] memberStorage = { INTEGERSIZE, MAXSTRINGSIZE, DOUBLESIZE, DOUBLESIZE };
 
         // Data size is the storage size for the members not the object.
-        static int getTotalDataSize() {
-            int data_size = 0;
+        static long getTotalDataSize() {
+            long data_size = 0;
             for (int indx = 0; indx < numberMembers; indx++)
                 data_size += memberStorage[indx] * memberDims[indx];
             return DIM0 * data_size;
         }
 
-        static int getDataSize() {
-            int data_size = 0;
+        static long getDataSize() {
+            long data_size = 0;
             for (int indx = 0; indx < numberMembers; indx++)
                 data_size += memberStorage[indx] * memberDims[indx];
             return data_size;
@@ -234,11 +234,11 @@ public class H5Ex_T_CompoundAttribute {
         }
 
         // Write the compound data.
-        dset_data = new byte[(int) dims[0] * Sensor_Datatype.getDataSize()];
+        dset_data = new byte[(int) dims[0] * (int)Sensor_Datatype.getDataSize()];
         ByteBuffer outBuf = ByteBuffer.wrap(dset_data);
         outBuf.order(ByteOrder.nativeOrder());
         for (int indx = 0; indx < (int) dims[0]; indx++) {
-            object_data[indx].writeBuffer(outBuf, indx * Sensor_Datatype.getDataSize());
+            object_data[indx].writeBuffer(outBuf, indx * (int)Sensor_Datatype.getDataSize());
         }
         try {
             if ((attribute_id >= 0) && (memtype_id >= 0))
@@ -395,7 +395,7 @@ public class H5Ex_T_CompoundAttribute {
         }
 
         // allocate memory for read buffer.
-        dset_data = new byte[(int) dims[0] * Sensor_Datatype.getDataSize()];
+        dset_data = new byte[(int) dims[0] * (int)Sensor_Datatype.getDataSize()];
 
         object_data2 = new Sensor[(int) dims[0]];
 
@@ -407,7 +407,7 @@ public class H5Ex_T_CompoundAttribute {
             ByteBuffer inBuf = ByteBuffer.wrap(dset_data);
             inBuf.order(ByteOrder.nativeOrder());
             for (int indx = 0; indx < (int) dims[0]; indx++) {
-                object_data2[indx] = new Sensor(inBuf, indx * Sensor_Datatype.getDataSize());
+                object_data2[indx] = new Sensor(inBuf, indx * (int)Sensor_Datatype.getDataSize());
             }
         }
         catch (Exception e) {
